attributeList parsing
Updates the lexer to include comments for attribute list parsing,
since we need to support commented-out JSON.

Implements JSON handling in the visitor by using JSON.parse()
after stripping out the comments.

Errors are adapted from the local JSON space to the overall
document space in a nasty function, but it's fully tested.
Using JSON.parse() saves us from needing to implement JSON
parsing and fully test it, but it comes at the cost of some
unnatural code in the parser. JSON errors will mostly only
be caught at the visitor stage, even what would normally be
lexer or syntax errors. Meanwhile, the visitor needs access
to the full document, so I added it using the custom payload
in the token. Since JS shares string references, this shouldn't
cause copies to be made and shouldn't exhaust memory on big
documents. If needed, the visitor could be provided with
the document directly instead.