Implement nested parsers & improve custom token providing mechanism
- Discovered chevrotain's token "category" feature, which allows us to
use a set of base "abstract" tokens for comments, etc. while the user
can provide any number of custom tokens for their language of choice.
- Convenience functions are provided to ensure that custom tokens
match the correct category, etc.
- Token payloads are used to ensure token pair matching (start/end token)
in the actual parser.
- Introduces the "PushParser" and "PopParser" token concept, which allows
you to push and pop another parser. For example, in markdown, three
backticks means a code block for a given language. This can be implemented
by pushing the other language's parser when you encounter the three
backticks. It actually keeps the main parser much simpler. In effect,
when it discovers a push parser token, the parser skips to the matching
pop parser token and treats the unparsed inner block of text as a
single pushParser rule. When the visitor goes to handle that rule, it
gets the parser details out of the given PushParser token payload
and passes the unparsed text to the other parser. The visitor then
merges the result with the current set of commands and continues.