Revert "Implement string literals"
This reverts commit 932f293e8cd58922f6374359f66a9991c6163df9.

Using the lexer to alternate modes was a flawed approach, as it would
be overeager in identifying string literals based on initial tokens
and not be able to revert back out of the mode unless it was a well-formed
literal (i.e. found a closing token), so error handling would break.
This might be ok if we were implementing a programming language,
but Bluehawk doesn't necessarily need to be strict about some tokens.