New Parser (#8)
* Implement lexer

* Parser WIP

* WIP

* CstVisitor WIP

* Mismatched command names

* WIP

* Fix empty command block with attribute bug

* Grammar update

* Share tokens between lexer and parser

* Nested comment blocks

* Parser updates

* CST Visitor WIP

* Update parser to handle many comments before command end

* Fully implement lineComment and blockComment context

* Semantic error reporting WIP

Fix asserts

* initial refactor of fileHandler to ts

* correctly handle file extensions

* Block and content ranges for commands (#9)

* Np constants (#11)

* constants js -> ts

* quiet all output

* Refactor get code block (#12)

* allow the latest hipster goodness

* refactor to ts, get rid of prop parsing

* attributeList parsing

Updates the lexer to include comments for attribute list parsing,
since we need to support commented-out JSON.

Implements JSON handling in the visitor by using JSON.parse()
after stripping out the comments.

Errors are adapted from the local JSON space to the overall
document space in a nasty function, but it's fully tested.
Using JSON.parse() saves us from needing to implement JSON
parsing and fully test it, but it comes at the cost of some
unnatural code in the parser. JSON errors will mostly only
be caught at the visitor stage, even what would normally be
lexer or syntax errors. Meanwhile, the visitor needs access
to the full document, so I added it using the custom payload
in the token. Since JS shares string references, this shouldn't
cause copies to be made and shouldn't exhaust memory on big
documents. If needed, the visitor could be provided with
the document directly instead.

* Step generator to ts (#13)

* early preparation for all message handling

* refactor to ts

* file updates removing legacy non jest testing

* refactor output.ts to messageHandler.ts

* Column always starts at 1

* Implement nested parsers & improve custom token providing mechanism

- Discovered chevrotain's token "category" feature, which allows us to
use a set of base "abstract" tokens for comments, etc. while the user
can provide any number of custom tokens for their language of choice.
- Convenience functions are provided to ensure that custom tokens
match the correct category, etc.
- Token payloads are used to ensure token pair matching (start/end token)
in the actual parser.
- Introduces the "PushParser" and "PopParser" token concept, which allows
you to push and pop another parser. For example, in markdown, three
backticks means a code block for a given language. This can be implemented
by pushing the other language's parser when you encounter the three
backticks. It actually keeps the main parser much simpler. In effect,
when it discovers a push parser token, the parser skips to the matching
pop parser token and treats the unparsed inner block of text as a
single pushParser rule. When the visitor goes to handle that rule, it
gets the parser details out of the given PushParser token payload
and passes the unparsed text to the other parser. The visitor then
merges the result with the current set of commands and continues.

* Validation (#14)

* Commandid-> range map (#15)

* Implement string literals

* Stitching together lexer-parser-visitor-validator (#16)

* Fix broken import

* Madness (#17)

* new command and processor systems

* remove needless property

Co-authored-by: Nathan Leniz <nathan.leniz@mongodb.com>

* var cleanup

* add jest coverage default

* Madness (#18)

* new command and processor systems

* remove needless property

* tests pass

* event plumbing and todo note

* big refactor

* eslint cleanup and test coverage

* recursive file input, state command uncomments

Co-authored-by: Nathan Leniz <nathan.leniz@mongodb.com>

* Commands transform (#19)

* Revert "Implement string literals"

This reverts commit 932f293e8cd58922f6374359f66a9991c6163df9.

Using the lexer to alternate modes was a flawed approach, as it would
be overeager in identifying string literals based on initial tokens
and not be able to revert back out of the mode unless it was a well-formed
literal (i.e. found a closing token), so error handling would break.
This might be ok if we were implementing a programming language,
but Bluehawk doesn't necessarily need to be strict about some tokens.

* Commands as transform implementation WIP

* Support end-to-end snippets test (#20)

- Add "--snippets" flag to cli to output snippets at destination
- Refactor BluehawkSource
- Fix slice -> snip
- Fix line command validator
- Support dedent on realistic codeblocks (not fully indented on blank lines)
- Protect against mismatch start/end tags
- Report errors on cli

* Support state output (#21)

* Support state output

* Add additional test for magic string assumptions

* Update README

* Update jest.config.js

Co-authored-by: Nathan <nathan.leniz@10gen.com>

* Update readme

Co-authored-by: Nathan <terakilobyte@gmail.com>
Co-authored-by: Nathan <nathan.leniz@10gen.com>
Co-authored-by: nate contino <nathan.contino@mongodb.com>
Co-authored-by: Nathan Leniz <nathan.leniz@mongodb.com>